requirements.txt
feedparser==6.0.11
prettytable==3.9.0
requests==2.31.0
openai-whisper @ git+https://github.com/openai/whisper.git
